+++ R/geom_fit_text.R
@@ -13,10 +13,11 @@
 #' There are three ways to define the box in which you want the text to be
 #' drawn. The extents of the box on the x and y axes are independent, so any
 #' combination of these methods can be used:
-#' 
+#'
 #' 1. If the `x` and/or `y` aesthetics are used to set the location of the box,
-#' the width or height will be set automatically based on the number of
-#' discrete values in `x` and/and `y`.
+#' the width and/or height will be set automatically based on the spacing
+#' between values on the axis, following the method used by
+#' `ggplot2::geom_boxplot()`.
 #' 2. Alternatively, if `x` and/or `y` aesthetics are used, the width and/or
 #' height of the box can be overridden with a 'width' and/or 'height' argument.
 #' These should be `grid::unit()` objects; if not, they will be assumed to use
@@ -46,15 +47,21 @@
 #' - angle
 #' - colour
 #' - family
+#' - fill (with `contrast = TRUE`, used to select a contrasting text colour)
 #' - fontface
 #' - lineheight
-#' - size
+#' - size (in points; note that this differs from `ggplot2::geom_text()`,
+#' which expresses size in mm)
+#'
+#' `geom_bar_text()` requires only the `x` and `y` aesthetics, which are used
+#' to infer the extent of the bar. If no `label` aesthetic is provided, the
+#' bar's value (`y`, or `x` for flipped bars) will be used as the label.
 #'
 #' @param padding.x,padding.y Horizontal and vertical padding around the text,
 #' expressed in `grid::unit()` objects. Both default to 1 mm.
-#' @param min.size Minimum font size, in points. Text that would need to be
-#' shrunk below this size to fit the box will be hidden. Defaults to 4 pt (8 pt
-#' for `geom_bar_text()`)
+#' @param min.size Minimum font size, in points. Text smaller than this, whether
+#' as supplied or after being shrunk to fit the box, will be hidden. Defaults to
+#' 4 pt (8 pt for `geom_bar_text()`).
 #' @param place Where inside the box to place the text. Default is 'centre';
 #' other options are 'topleft', 'top', 'topright', 'right', 'bottomright',
 #' 'bottom', 'bottomleft', 'left', and 'center'/'middle' which are both
@@ -63,24 +70,26 @@
 #' @param outside If `TRUE`, text placed in one of 'top', 'right', 'bottom' or
 #' 'left' that would need to be shrunk smaller than `min.size` to fit the box
 #' will be drawn outside the box if possible. This is mostly useful for drawing
-#' text inside bar/column geoms. Defaults to TRUE for `position = "identity"`
-#' when using `geom_bar_text()`, otherwise FALSE.
+#' text inside bar/column geoms. Defaults to `TRUE` for `position = "identity"`
+#' when using `geom_bar_text()`, otherwise `FALSE`.
 #' @param grow If `TRUE`, text will be grown as well as shrunk to fill the box.
-#' Defaults to FALSE.
+#' Defaults to `FALSE`.
 #' @param reflow If `TRUE`, text will be reflowed (wrapped) to better fit the
-#' box. Defaults to FALSE.
-#' @param hjust,vjust Horizontal and vertical justification of the text. By
-#' default, these are automatically set to appropriate values based on `place`.
+#' box. Defaults to `FALSE`.
+#' @param hjust,vjust Horizontal and vertical justification of the text.
+#' `vjust` defaults to 0.5. `hjust` defaults to 0.5, except for left placements
+#' ('left', 'topleft' and 'bottomleft'), where it defaults to 0, and right
+#' placements, where it defaults to 1.
 #' @param fullheight If `TRUE`, descenders will be counted when resizing and
 #' placing text; if `FALSE`, only the x-height and ascenders will be counted.
 #' The main use for this option is for aligning text at the baseline (`FALSE`)
 #' or preventing descenders from spilling outside the box (`TRUE`). By default
-#' this is set automatically depending on `place` and `grow`. 
+#' this is set automatically depending on `grow`.
 #' @param width,height When using `x` and/or `y` aesthetics, these set the
 #' width and/or height of the box. These should be either `grid::unit()`
 #' objects or numeric values on the `x` and `y` scales.
 #' @param formatter A function that will be applied to the text before it is
-#' drawn. This is useful when using `geom_fit_text()` in context involving
+#' drawn. This is useful when using `geom_fit_text()` in contexts involving
 #' interpolated variables, such as with the 'gganimate' package. `formatter`
 #' will be applied serially to each element in the `label` column, so it does
 #' not need to be a vectorised function.
@@ -93,7 +102,7 @@
 #' @param rich If `TRUE`, text will be formatted with markdown and HTML markup
 #' as implemented by `gridtext::richtext_grob()`. `FALSE` by default. Rich text
 #' cannot be drawn in polar coordinates. Please note that rich text support is
-#' **experimental** and breaking changes are likely
+#' **experimental** and breaking changes are likely.
 #' @param data,stat,position,na.rm,show.legend,inherit.aes,... Standard geom
 #' arguments as for `ggplot2::geom_text()`.
 #' @param flip If `TRUE`, when in polar coordinates 'upside-down' text will be
@@ -188,9 +197,12 @@ GeomFitText <- ggplot2::ggproto(
   ),
 
   setup_params = function(data, params) {
-
     # Standardise the place argument
-    params$place <- ifelse(params$place %in% c("middle", "center"), "centre", params$place)
+    params$place <- ifelse(
+      params$place %in% c("middle", "center"),
+      "centre",
+      params$place
+    )
 
     params
   },
@@ -199,25 +211,24 @@ GeomFitText <- ggplot2::ggproto(
     data,
     params
   ) {
-
     # Check that valid aesthetics have been supplied for each dimension
-    if (! (! is.null(data$xmin) & ! is.null(data$xmax) | ! is.null(data$x))) {
+    if (!(!is.null(data$xmin) & !is.null(data$xmax) | !is.null(data$x))) {
       cli::cli_abort("geom_fit_text needs either 'xmin' and 'xmax', or 'x'")
     }
-    if (! (! is.null(data$ymin) & ! is.null(data$ymax) | ! is.null(data$y))) {
+    if (!(!is.null(data$ymin) & !is.null(data$ymax) | !is.null(data$y))) {
       cli::cli_abort("geom_fit_text needs either 'ymin' and 'ymax', or 'y'")
     }
 
     # If 'width' is provided, but not as unit, interpret it as a numeric on the
     # x scale
-    if ((! is.null(params$width)) & (! inherits(params$width, "unit"))) {
+    if ((!is.null(params$width)) & (!inherits(params$width, "unit"))) {
       data$xmin <- data$x - params$width / 2
       data$xmax <- data$x + params$width / 2
     }
 
     # If 'height' is provided, but not a unit, interpret it as a numeric on the
     # y scale
-    if ((! is.null(params$height)) & (! inherits(params$height, "unit"))) {
+    if ((!is.null(params$height)) & (!inherits(params$height, "unit"))) {
       data$ymin <- data$y - params$height / 2
       data$ymax <- data$y + params$height / 2
     }
@@ -241,24 +252,30 @@ GeomFitText <- ggplot2::ggproto(
     }
 
     # Apply a formatter function, if one was given
-    if (! is.null(params$formatter)) {
-
+    if (!is.null(params$formatter)) {
       # Check that 'formatter' is a function
-      if (! is.function(params$formatter)) {
+      if (!is.function(params$formatter)) {
         cli::cli_abort("`formatter` must be a function")
       }
 
       # Apply formatter to the labels, checking that the output is a character
       # vector of the correct length
-      formatted_labels <- vapply(data$label, params$formatter, character(1), USE.NAMES = FALSE)
-      if ((! length(formatted_labels) == length(data$label)) | 
-          (! is.character(formatted_labels))) {
-        cli::cli_abort("`formatter` must produce a character vector of same length as input")
+      formatted_labels <- vapply(
+        data$label,
+        params$formatter,
+        character(1),
+        USE.NAMES = FALSE
+      )
+      if (
+        (!length(formatted_labels) == length(data$label)) |
+          (!is.character(formatted_labels))
+      ) {
+        cli::cli_abort(
+          "`formatter` must produce a character vector of same length as input"
+        )
       }
       data$label <- formatted_labels
     }
-
-    data$flip <- params$flip
 
     data
   },
@@ -286,7 +303,6 @@ GeomFitText <- ggplot2::ggproto(
     rich = FALSE,
     flip = FALSE
   ) {
-
     # Rich text cannot be used in polar coordinates
     if (inherits(coord, "CoordPolar") & rich) {
       cli::cli_abort("Cannot draw rich text in polar coordinates")
@@ -295,7 +311,9 @@ GeomFitText <- ggplot2::ggproto(
     # Transform data to plot scales; if in polar coordinates, we need to ensure
     # that x and y values are given
     if (inherits(coord, "CoordPolar")) {
-      if (is.null(data$x)) data$x <- 1
+      if (is.null(data$x)) {
+        data$x <- 1
+      }
       if (is.null(data$y)) data$y <- 1
     }
     data <- coord$transform(data, panel_scales)
@@ -304,16 +322,16 @@ GeomFitText <- ggplot2::ggproto(
     # theta and r values respectively; these can be used later to accurately
     # set the width and height of the bounding box in polar space
     if (inherits(coord, "CoordPolar")) {
-      if (! is.null(data$xmin)) {
+      if (!is.null(data$xmin)) {
         data$xmin <- theta_rescale(coord, data$xmin, panel_scales)
       }
-      if (! is.null(data$xmax)) {
+      if (!is.null(data$xmax)) {
         data$xmax <- theta_rescale(coord, data$xmax, panel_scales)
       }
-      if (! is.null(data$ymin)) {
+      if (!is.null(data$ymin)) {
         data$ymin <- r_rescale(coord, data$ymin, panel_scales$r.range)
       }
-      if (! is.null(data$ymax)) {
+      if (!is.null(data$ymax)) {
         data$ymax <- r_rescale(coord, data$ymax, panel_scales$r.range)
       }
     }
@@ -335,7 +353,11 @@ GeomFitText <- ggplot2::ggproto(
       contrast = contrast,
       rich = rich,
       flip = flip,
-      cl = ifelse(inherits(coord, "CoordPolar"), "fittexttreepolar", "fittexttree")
+      cl = ifelse(
+        inherits(coord, "CoordPolar"),
+        "fittexttreepolar",
+        "fittexttree"
+      )
     )
     gt$name <- grid::grobName(gt, "geom_fit_text")
     gt
@@ -345,7 +367,6 @@ GeomFitText <- ggplot2::ggproto(
 #' @importFrom grid makeContent
 #' @export
 makeContent.fittexttree <- function(x) {
-
   # Extract data
   ftt <- x
   data <- ftt$data
@@ -387,15 +408,18 @@ makeContent.fittexttree <- function(x) {
   }
 
   # Remove any rows with NA boundaries
-  na_rows <- which(is.na(data$xmin) | is.na(data$xmax) | is.na(data$ymin) | 
-                   is.na(data$ymax))
+  na_rows <- which(
+    is.na(data$xmin) | is.na(data$xmax) | is.na(data$ymin) | is.na(data$ymax)
+  )
   if (length(na_rows) > 0) {
     data <- data[-na_rows, ]
-    cli::cli_warn("Removed {length(na_rows)} rows where box limits were outside plot limits ({.fun geom_fit_text}).")
+    cli::cli_warn(
+      "Removed {length(na_rows)} rows where box limits were outside plot limits ({.fun geom_fit_text})."
+    )
   }
 
   # Remove any rows with blank labels
-  data <- data[which(! is.na(data$label) | data$label == ""), ]
+  data <- data[!is.na(data$label) & data$label != "", ]
 
   # Clean up angles
   data$angle <- data$angle %% 360
@@ -403,10 +427,9 @@ makeContent.fittexttree <- function(x) {
   # If contrast is set, and the shade of a text colour is too close to the
   # shade of the fill colour, change the colour to its complement
   if (ftt$contrast) {
-
     # If any fill value is NA, emit a warning and set to the default
     # ggplot2 background grey
-    if (! is.null(data$fill)) {
+    if (!is.null(data$fill)) {
       if (any(is.na(data$fill))) {
         cli::cli_warn("NA values in fill")
         data$fill <- data$fill %NA% "grey35"
@@ -433,7 +456,6 @@ makeContent.fittexttree <- function(x) {
 
   # Prepare grob for each text label
   grobs <- lapply(seq_len(nrow(data)), function(i) {
-
     # Convenience
     text <- data[i, ]
 
@@ -442,7 +464,15 @@ makeContent.fittexttree <- function(x) {
     ydim <- hnpc2mm(abs(text$ymin - text$ymax) - (2 * ftt$padding.y))
 
     # Reflow and/or resize the text into a textGrob
-    tg <- reflow_and_resize(text, ftt$reflow, ftt$grow, ftt$fullheight, xdim, ydim, ftt$rich)
+    tg <- reflow_and_resize(
+      text,
+      ftt$reflow,
+      ftt$grow,
+      ftt$fullheight,
+      xdim,
+      ydim,
+      ftt$rich
+    )
 
     # If the font size is too small and 'outside' has been set, try reflowing
     # and resizing again in the 'outside' position
@@ -474,17 +504,28 @@ makeContent.fittexttree <- function(x) {
         bg_colour <- "grey92"
         text_colour <- text$colour %||% "black"
         if (
-          abs(shades::lightness(bg_colour) - shades::lightness(text_colour)) < 50
+          abs(shades::lightness(bg_colour) - shades::lightness(text_colour)) <
+            50
         ) {
           complement <- shades::complement(shades::shade(text_colour))
           text$colour <- as.character(complement)
         }
       }
-      tg <- reflow_and_resize(text, ftt$reflow, ftt$grow, ftt$fullheight, xdim, ydim, ftt$rich)
+      tg <- reflow_and_resize(
+        text,
+        ftt$reflow,
+        ftt$grow,
+        ftt$fullheight,
+        xdim,
+        ydim,
+        ftt$rich
+      )
     }
 
     # If the font size is still too small, don't draw this label
-    if (getfontsize(tg) < ftt$min.size) return()
+    if (getfontsize(tg) < ftt$min.size) {
+      return()
+    }
 
     # Set hjust and vjust
     tg$hjust <- ftt$hjust
@@ -518,13 +559,13 @@ makeContent.fittexttree <- function(x) {
     # We can use these values to calculate the magnitude of the vector from the
     # centre point to the anchor point, using the Pythagorean identity
     if (ftt$fullheight) {
-      rise <- ((tg_height_unrot * tg$vjust) + tg_descent_unrot) - 
-                ((tg_height_unrot + tg_descent_unrot) * 0.5)
+      rise <- ((tg_height_unrot * tg$vjust) + tg_descent_unrot) -
+        ((tg_height_unrot + tg_descent_unrot) * 0.5)
     } else {
       rise <- (tg_height_unrot * tg$vjust) - (tg_height_unrot * 0.5)
     }
     run <- (tg_width_unrot * tg$hjust) - (tg_width_unrot * 0.5)
-    magnitude <- sqrt((rise ^ 2) + (run ^ 2))
+    magnitude <- sqrt((rise^2) + (run^2))
 
     # The angle between the baseline and the vector can be calculated from the
     # known rise and run
@@ -596,24 +637,19 @@ makeContent.fittexttree <- function(x) {
   grid::setChildren(ftt, grobs)
 }
 
-#' geom_fit_text
-#' @noRd
-geom_grow_text <- function(...) {
-  .Deprecated("geom_fit_text(grow = T, ...)")
-}
-
-#' geom_fit_text
-#' @noRd
-geom_shrink_text <- function(...) {
-  .Deprecated("geom_fit_text(grow = F, ...)")
-}
-
 #' Return a textGrob with the text reflowed and/or resized to fit given
 #' dimensions
 #'
 #' @noRd
-reflow_and_resize <- function(text, reflow, grow, fullheight, xdim, ydim, rich) {
-
+reflow_and_resize <- function(
+  text,
+  reflow,
+  grow,
+  fullheight,
+  xdim,
+  ydim,
+  rich
+) {
   # Create textGrob or richtext_grob
   if (rich) {
     tg <- gridtext::richtext_grob(
@@ -633,11 +669,20 @@ reflow_and_resize <- function(text, reflow, grow, fullheight, xdim, ydim, rich) 
       ),
       use_markdown = TRUE
     )
-    tg$params <- list(text = text$label, x = 0.5, y = 0.5, hjust = 0.5, vjust = 0.5, 
-                      rot = text$angle, colour = text$colour, alpha = text$alpha, 
-                      fontsize = text$size, fontfamily = text$family, 
-                      fontface = text$fontface, lineheight = text$lineheight)
-
+    tg$params <- list(
+      text = text$label,
+      x = 0.5,
+      y = 0.5,
+      hjust = 0.5,
+      vjust = 0.5,
+      rot = text$angle,
+      colour = text$colour,
+      alpha = text$alpha,
+      fontsize = text$size,
+      fontfamily = text$family,
+      fontface = text$fontface,
+      lineheight = text$lineheight
+    )
   } else {
     tg <- grid::textGrob(
       label = text$label,
@@ -662,9 +707,11 @@ reflow_and_resize <- function(text, reflow, grow, fullheight, xdim, ydim, rich) 
 
   # Reflow the text, if reflow = TRUE and either the text doesn't currently
   # fit or grow = TRUE and if text contains spaces
-  if (reflow & (grow | tgdim$width > xdim | tgdim$height > ydim) & 
-      stringi::stri_detect_regex(getlabel(tg), "\\s")) {
-
+  if (
+    reflow &
+      (grow | tgdim$width > xdim | tgdim$height > ydim) &
+      stringi::stri_detect_regex(getlabel(tg), "\\s")
+  ) {
     # Generate a list of all wraps for the text
     wraps <- wraplabel(tg)
 
@@ -677,19 +724,17 @@ reflow_and_resize <- function(text, reflow, grow, fullheight, xdim, ydim, rich) 
     )
     wraps$aspect_ratio <- vapply(
       wraps$tgdim,
-      function(tgdim) tgdim$width/tgdim$height,
+      function(tgdim) tgdim$width / tgdim$height,
       double(1)
     )
     wraps$ar_diff <- abs(wraps$aspect_ratio - (xdim / ydim))
 
     # If grow is false, select the widest wrap that fits the box (if there is
     # one) and return it
-    if (! grow) {
-
+    if (!grow) {
       wraps$width <- unlist(lapply(wraps$tgdim, function(tgdim) tgdim$width))
       wraps$height <- unlist(lapply(wraps$tgdim, function(tgdim) tgdim$height))
-      fit_wraps <- wraps[which(wraps$width < xdim), ]
-      fit_wraps <- wraps[which(fit_wraps$height < ydim), ]
+      fit_wraps <- wraps[wraps$width < xdim & wraps$height < ydim, ]
 
       if (nrow(fit_wraps) > 0) {
         best_index <- which(fit_wraps$wrapwidth == max(fit_wraps$wrapwidth))[1]
@@ -713,10 +758,9 @@ reflow_and_resize <- function(text, reflow, grow, fullheight, xdim, ydim, rich) 
   if (
     # Standard condition - is text too big for box?
     (tgdim$width > xdim | tgdim$height > ydim) |
-    # grow = TRUE condition - is text too small for box?
-    (grow & tgdim$width < xdim & tgdim$height < ydim)
+      # grow = TRUE condition - is text too small for box?
+      (grow & tgdim$width < xdim & tgdim$height < ydim)
   ) {
-
     # Get the slopes of the relationships between font size and label
     # dimensions
     slopew <- getfontsize(tg) / tgdim$width